Change ‘SYNOPSIS’ to ‘DESCRIPTION’ and make usage output verbose
Normally Perl programs have a SYNOPSIS section in the POD area.  This
patch changes the name because Pod::Usage shows both SYNOPSIS and
USAGE under the name ‘Usage’, which looks slightly confusion to me.  I
would rather break the tradition and name the section DESCRIPTION so
that it has a unique name when the user sees help output from running
the program.

This patch also places the call to pod2usage() inside of a function to
keep the main “show help and exit if needed” logic a nice one-liner.