Change Makefile.PL to store my repository as additional meta info
Squeeks uses his own URI for the fork of the program when providing
extra meta data.  This patch changes the URI to my repository.  This
is not an attempt to belittle him or dimish the value of his
contributions, but for the time being I expect more expects to may
copy than his.